LayoutTests/storage/indexeddb/deletedatabase-delayed-by-open-and-versionchange-expected.txt

Test that deleteDatabase is delayed if a VERSION_CHANGE transaction is running

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "deletedatabase-delayed-by-open-and-versionchange.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)

prepareDatabase():
versionChangeComplete = false
h = event.target.result
indexedDB.deleteDatabase(dbname)

onVersionChange():
old = 1
new = ""

deleteDatabaseOnBlocked():

transactionOnComplete():
versionChangeComplete = true

onOpenSuccess():
h = event.target.result
h.close()

deleteDatabaseOnSuccess():
PASS versionChangeComplete is true
PASS successfullyParsed is true

TEST COMPLETE
